repo_url in Project model by #71
