+++ ui/templates/project_settings.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html>
 <body>
-<form action="/ui/{{projectname}}/settings" method="post">
+<form action='/{{projectname}}/settings' method='post'>
     {% csrf_token %}
     {{ form }}</br>
-    <input type="submit" value="Save"/>
+    <input type='submit' value='Save'/>
 </form>
 </body>
 </html>
